Move ElementIterator and SymbolsIterator back to `elements` module
